+++ src/Request.php
@@ -12,7 +12,7 @@ use Busarm\PhpMini\Session\PHPSession;
 use Busarm\PhpMini\Interfaces\Bags\AttributeBag;
 use Busarm\PhpMini\Interfaces\Bags\SessionBag;
 use Busarm\PhpMini\Interfaces\RequestInterface;
-use Closure;
+use Busarm\PhpMini\Traits\Container;
 use Psr\Http\Message\ServerRequestInterface;
 use Psr\Http\Message\UriInterface;
 
@@ -34,6 +34,8 @@ use function Busarm\PhpMini\Helpers\is_https;
  */
 class Request implements RequestInterface
 {
+    use Container;
+    
     protected string|null $ip           =   NULL;
     protected string|null $scheme       =   NULL;
     protected string|null $domain       =   NULL;
@@ -56,7 +58,6 @@ class Request implements RequestInterface
 
     private array $sessionOptions   =   [];
     private array $cookieOptions    =   [];
-    private array $singletons       =   [];
 
     /**
      * Constructor.
@@ -600,30 +601,5 @@ class Request implements RequestInterface
         $this->headers = $headers;
 
         return $this;
-    }
-
-    /**
-     * Add singleton
-     * 
-     * @param string $className
-     * @param object|null $object
-     * @return self
-     */
-    public function addSingleton(string $className, &$object)
-    {
-        $this->singletons[$className] = $object;
-        return $this;
-    }
-
-    /**
-     * Get singleton
-     *
-     * @param string $className
-     * @param object $default
-     * @return self
-     */
-    public function getSingleton(string $className, $default = null)
-    {
-        return $this->singletons[$className] ?? $default;
     }
 }