+++ CHANGELOG.md
@@ -2,6 +2,17 @@
 
 All notable changes to this project will be documented in this file.
 
+## [1.1.7]
+
+### Fixed
+
+-   Dependency Injection for bug in constructors for request and response object
+
+### Added
+
+-   Container Interface and Trait
+-   Routes, support singletons
+
 ## [1.1.6]
 
 ### Changed